requirements.txt
# Core
Django==5.0.2
djangorestframework==3.14.0

# CORS
django-cors-headers==4.3.1

# Environment variables
python-dotenv==1.0.1

# Database (PostgreSQL for production)
psycopg2-binary==2.9.9

# API Documentation
drf-spectacular==0.27.1

# Utilities
Pillow==10.2.0
python-dateutil==2.8.2
